feat(api): decouple data transformers from API layer
